pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "django>=5.1.3",
    "django-cors-headers>=4.6.0",
    "django-crontab>=0.7.1",
    "django-rest-framework>=0.1.0",
    "getmac>=0.9.5",
    "matplotlib>=3.9.3",
    "networkx>=3.4.2",
    "pandas>=2.2.3",
    "python-nmap>=0.7.1",
    "requests>=2.32.3",
    "scapy>=2.6.1",
    "speedtest-cli>=2.1.3",
    "torch>=2.5.1",
    "traffic>=2.11",
    "transformers>=4.46.3",
    "werkzeug>=3.1.3",
]
